lab3/pyproject.toml
[project]
name = "sensor-emulator"
version = "0.1.0"
description = ""
authors = [
    {name = "volodymyrd2",email = "volodymyrd@jobtestprep.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "google-cloud-pubsub (>=2.33.0,<3.0.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "sensor_emulator", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
